01-docker-terraform/requirements.txt
# Core dependencies
pandas>=2.0.0,<3.0.0
pyarrow>=14.0.0,<15.0.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
numpy==1.26.4

# CLI
click>=8.1.0,<9.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0

# Jupyter (for notebooks)
jupyter>=1.0.0,<2.0.0
ipykernel>=6.25.0,<7.0.0
